+++ seed/options.js
@@ -1,23 +1,29 @@
-module.exports = {
-    framework: [
+module.exports = [{
+    message: 'Library / Framework?',
+    name: 'framework',
+
+    choices: [
         { name: 'Polymer', value: 'polymer' }
-    ],
-
-    polymer: {
-        compiler: [
-            { name: 'BabelJS', value: 'babeljs' }
-        ],
-
-        css: [
-            { name: 'SaSS', value: 'sass' }
-        ],
-
-        html: [
-            { name: 'Jade', value: 'jade' }
-        ]
-    },
-
-    test: [
-        { name: 'Mocha', value: 'mocha' }
     ]
-}
+}, {
+    message: 'HTML?',
+    name: 'html',
+
+    choices: [
+        { name: 'Jade', value: 'jade' }
+    ]
+}, {
+    message: 'JS?',
+    name: 'js',
+
+    choices: [
+        { name: 'BabelJS', value: 'babeljs', extra: 'js' }
+    ]
+}, {
+    message: 'CSS?',
+    name: 'css',
+
+    choices: [
+        { name: 'SaSS', value: 'sass', extra: 'sass' }
+    ]
+}]