+++ seed/template/all/root-modified-by-user/eslintrc.js
@@ -1,9 +1,8 @@
-module.exports = () => {
+module.exports = ($) => {
     const speedseed = require('speedseed')
-
     const file = new speedseed.Files()
 
-    file.updateFile('.eslintrc', 4, {
+    const optionsGeneral = {
         // 'extends': 'eslint:recommended',
         env: {
             es6: true
@@ -16,14 +15,14 @@ module.exports = () => {
             // POSSIBLE ERRORS
             'comma-dangle': 2, // RECOMMENDED
             'no-cond-assign': 2, // RECOMMENDED
-            'no-console': 1, // RECOMMENDED
+            'no-console': 0, // RECOMMENDED
             'no-constant-condition': 2, // RECOMMENDED
             'no-control-regex': 2, // RECOMMENDED
             'no-debugger': 1, // RECOMMENDED
             'no-dupe-args': 2, // RECOMMENDED
             'no-dupe-keys': 2, // RECOMMENDED
             'no-duplicate-case': 2, // RECOMMENDED
-            'no-empty': 2, // RECOMMENDED
+            'no-empty': 0, // RECOMMENDED
             'no-empty-character-class': 2, // RECOMMENDED
             'no-ex-assign': 2, // RECOMMENDED
             'no-extra-boolean-cast': 2, // RECOMMENDED
@@ -47,9 +46,13 @@ module.exports = () => {
             'accessor-pairs': 0,
             'array-callback-return': 2,
             'block-scoped-var': 2,
-            'complexity': [1, 2],
+            'complexity': [1, 6],
             'consistent-return': 0,
-            'curly': 2
+            'curly': 0
         }
-    })
+    }
+
+    file.extendFromUser(optionsGeneral, './.eslintrc')
+
+    file.writeFile('.eslintrc', optionsGeneral, 4)
 }