+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 - [Spanish](#spanish)
 
 ## English
+### v.0.3.0
+    - Updated to Polymer v.1.7.0
+    - Updated to generator-speedseed v.0.22.0
+    - Updated to speedseed v.0.18.0
+
 ### v.0.2.1
     - Updated to generator-speedseed v.0.21.0
     - Updated to speedseed v.0.17.0
@@ -14,6 +19,11 @@
     - Init
 
 ## Spanish
+### v.0.3.0
+    - Actualizado a Polymer v.1.7.0
+    - Actualizado a generator-speedseed v.0.22.0
+    - Actualizado a speedseed v.0.18.0
+
 ### v.0.2.1
     - Actualizado a generator-speedseed v.0.21.0
     - Actualizado a speedseed v.0.17.0